+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [1.6.0] - 2019-02-23
+
+### Added
+- QUERY
+    -  createRetakeTest ( API to create Retake Test)
+    -  updateRetakeTest ( API to update Retake Test)
+    -  moveRetakeTest ( API to move Retake Test )
+
+### Changes
+    - Change TestForStudentProfile and CommonAnalysisForStudentProfile to get retake test list and analysis.
+
 ## [1.5.0] - 2019-02-11
 
 ### Added